static cast for #160

+++ src/flows.cpp
@@ -147,7 +147,7 @@ struct OneAggregate : public RcppParallel::Worker
                             while (target_t < INFINITE_INT)
                             {
                                 path_len++;
-                                target_t = prev [target_t];
+                                target_t = prev [static_cast <size_t> (target_t)];
                                 if (target_t < 0 || target_t == from_t)
                                     break;
                             }
